Put the code to prepend theme asset paths into a before_initialize block
A workaround for Rails 5 (will not break Rails 4.2 themes).

This allows the theme assets to be prepended to Sprockets' asset paths
(they need to be the first thing in the array to override the app paths)
after the call to `Rails.application.initialize!` otherwise Rails 5 will
add them too early, causing them to be shuffled to the end of the array
as everything else loads (prepends) afterwards.